refactor: merge deprecated statement checker into settings checker
Continue moving the linter towards a single AST visitor.
`DeprecatedStatementChecker` was already a pure dispatcher, and every node type
it visited (`SectionHeader`, `LibraryImport`, `ForceTags`, `Return`,
`ReturnSetting` and `Keyword`) is already visited by `SettingsChecker`. Its four
rules are now handled by `SettingsChecker`, which removes one more full walk
over every parsed file.

This reduces the number of AST visitors from 24 to 23. Linter output is
unchanged.
